Added an empty Visual Studio project.
Note that the target framework is .NET 3.5, in order to make this library
as reusable as possible. The reason for not going even further back is
that LINQ Expressions will be very handy for provding strongly typed
access to Reflection instances such as PropertyInfo, MethodInfo, etc.

However, as far as I can tell, none of the features added in .NET 4+
provides much benefit in this context.

However, this is a decision, which can still be debated if need be.