Added a Guard Clause to protect against
incompatible properties. The exception text is patterned off the BCL
exception text for fields.

The check for compatibility is based on
RuntimeMethodInfo.CheckConsistency, which is a private BCL method.
However, the check is simplified, so in the end, it may be necessary with
a try/catch block instead of the Guard Clause.